Refs #124. Moved the faked superscript/subscript handling to a mixin
